requirements.txt
boto3
botocore
paramiko
click
tabulate
pyyaml
